+++ RELEASE_NOTES.md
@@ -1,4 +1,11 @@
-Version 2.3.1
+KBaseReport Release Notes
+=========================================
 
-Bug fix to cleanup zip area.
+3.0.0
+-----
+- Refactored the app into Python with improved tests and documentation
+- Split the KIDL spec into KBaseReport.spec (for the API) and KBaseReportWorkspace (for the Workspace)
 
+2.3.1
+-----
+- Bug fix to cleanup zip area